+++ src/main/java/br/com/agendamento/api/dto/usuario/UsuarioCadastroDTO.java
@@ -6,14 +6,12 @@ import javax.validation.constraints.Size;
 
 public record UsuarioCadastroDTO(
         @NotBlank
-        @Size(max = 50, message = "Excedido limite de 50 caracteres")
+        @Size(max = 50, message = "Excedido limite de 50 caractere")
         String nome,
         @NotBlank
         @Email
         @Size(max = 100, message = "Excedido limite de 100 caracteres")
         String email,
-        @NotBlank
-        @Size(max = 100, message = "Excedido limite de 100 caracteres")
         String senha,
         @NotBlank
         String senhaConfirma
